Personalization/requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg2-binary